CRAFTER-1254: Unable to use tags as facets properly because the field doesn't end with _s
The data sources "configured list" and "static key value pairs" both now
support the data type field in their properties. The checkbox group has
been fixed to make use of this data type when saving values into the
model. Also, a new property control was created (dropdown) that allows
the user to only select from a list of options. The options can be
easily updated by changing the correspondng configuration (in
key-value-list.js or configured-list.js). After a checkbox group has
been assigned a data source, the user is still able to change data source;
though this may result in a loss of information if the options selected
do not exist in the new data source assigned.