Enhancements to the date time control
Fixes CRAFTER-1282 and CRAFTERCMS-10. For CRAFTER-1282, a clear link
was added so the user can clear the date and time values. For
CRAFTERCMS-10, a navigator was added to the calendar which opens when
the user clicks on the month/year. The navigator lets the user skip to
a specific month or year.